Clean up the code, migrate to sass and typescript, also convert file names and css classnames to lowercase
