--- src/components/Work/Project/Project.module.css
@@ -1,106 +1,0 @@
-.Project {
-    display: grid;
-    grid-template-columns: 2fr 1fr;
-    padding: 1.5rem;
-    border-radius: 5px;
-    box-shadow: 0 0 4px #bdbdbd;
-    transition: box-shadow 200ms ease;
-}
-
-.Project:hover {
-    box-shadow: 0 0 8px #bdbdbd;
-}
-
-.Project.Dark {
-    box-shadow: 0 0 4px #111111;
-}
-
-.Project.Dark:hover {
-    box-shadow: 0 0 8px #111111;
-}
-
-.DesktopPreview {
-    width: 100%;
-    height: 180px;
-    background-color: #bdbdbd;
-}
-
-.MobilePreview {
-    display: none;
-    width: 100%;
-    height: 180px;
-    background-color: #bdbdbd;
-}
-
-:is(.DesktopPreview, .MobilePreview) > img {
-    width: 100%;
-    height: 100%;
-}
-
-.Project h2 {
-    font-size: 2rem;
-}
-
-.Links {
-    margin: 1rem 0;
-}
-
-.Links > * {
-    margin: 0 0.2rem;
-}
-
-.Description {
-    max-width: 500px;
-    font-size: 1.1rem;
-    margin-right: 1rem;
-    margin-bottom: 1rem;
-    text-align: justify;
-}
-
-.Technologies > h3 {
-    font-size: 1.6rem;
-    margin: 1rem 0;
-}
-
-.Technologies > ul {
-    padding-left: 2rem;
-    font-size: 1.2rem;
-}
-
-.Technologies > ul > li {
-    color: #2c3842;
-    margin: 4px 0;
-    list-style: square;
-}
-
-/* Dark mode technologies li tags color */
-.Project.Dark .Technologies > ul > li {
-    color: #ececec;
-}
-
-@media screen and (max-width: 680px) {
-    .Project {
-        grid-template-columns: 1fr;
-    }
-
-    .Project h2 {
-        font-size: 1.5rem;
-    }
-
-    .DesktopPreview {
-        display: none;
-    }
-
-    .MobilePreview {
-        display: block;
-    }
-
-    .Links {
-        text-align: center;
-    }
-
-    .Description {
-        font-size: 1rem;
-        margin-right: 0;
-    }
-}
