--- src/components/Work/Work.module.css
@@ -1,19 +1,0 @@
-.Work {
-    padding: 1rem;
-}
-
-.WorkWrapper {
-    width: 100%;
-    max-width: 980px;
-    margin: 0 auto;
-}
-
-.WorkWrapper > h2 {
-    font-size: 2rem;
-    margin: 1rem 0;
-    text-align: center;
-}
-
-.Projects > * {
-    margin: 1.5rem 0;
-}
