+++ src/components/work/project/project.module.scss
@@ -1,0 +1,95 @@
+.project {
+  display: grid;
+  grid-template-columns: 2fr 1fr;
+  padding: 1.5rem;
+  border-radius: 5px;
+  box-shadow: 0 0 4px #bdbdbd;
+  transition: box-shadow 200ms ease;
+
+  h2 {
+    font-size: 2rem;
+  }
+
+  &:hover {
+    box-shadow: 0 0 8px #bdbdbd;
+  }
+}
+
+.desktopPreview {
+  width: 100%;
+  height: 180px;
+  background-color: #bdbdbd;
+}
+
+.mobilePreview {
+  display: none;
+  width: 100%;
+  height: 180px;
+  background-color: #bdbdbd;
+}
+
+:is(.desktopPreview, .mobilePreview) > img {
+  width: 100%;
+  height: 100%;
+}
+
+.links {
+  margin: 1rem 0;
+
+  & > * {
+    margin: 0 0.2rem;
+  }
+}
+
+.description {
+  max-width: 500px;
+  font-size: 1.1rem;
+  margin-right: 1rem;
+  margin-bottom: 1rem;
+  text-align: justify;
+}
+
+.technologies {
+  h3 {
+    font-size: 1.6rem;
+    margin: 1rem 0;
+  }
+
+  ul {
+    padding-left: 2rem;
+    font-size: 1.2rem;
+
+    li {
+      color: #2c3842;
+      margin: 4px 0;
+      list-style: square;
+    }
+  }
+}
+
+@media screen and (max-width: 680px) {
+    .project {
+      grid-template-columns: 1fr;
+
+      h2 {
+        font-size: 1.5rem;
+      }
+    }
+
+    .desktopPreview {
+      display: none;
+    }
+
+    .mobilePreview {
+      display: block;
+    }
+
+    .links {
+      text-align: center;
+    }
+
+    .description {
+      font-size: 1rem;
+      margin-right: 0;
+    }
+}
