+++ src/components/work/work.module.scss
@@ -1,0 +1,19 @@
+.work {
+  padding: 1rem;
+}
+
+.workWrapper {
+  width: 100%;
+  max-width: 980px;
+  margin: 0 auto;
+
+  h2 {
+    font-size: 2rem;
+    margin: 1rem 0;
+    text-align: center;
+  }
+}
+
+.projects > * {
+  margin: 1.5rem 0;
+}
