+++ src/pages/index.tsx
@@ -1,5 +1,5 @@
-import { Showcase } from '../components/Showcase/Showcase';
-import { Work } from '../components/Work/Work';
+import { Showcase } from '../components/showcase/showcase';
+import { Work } from '../components/work/work';
 import { Contact } from '../components/Contact/Contact';
 
 export default function Home() {
